Split file.
